Task to create Steam Workshop Item

+++ eawx-build/Steam/ISteamWorkshop.cs
@@ -4,10 +4,10 @@ namespace EawXBuild.Steam
 {
     public interface ISteamWorkshop
     {
-        int AppId { get; set; }
+        uint AppId { get; set; }
 
         Task<PublishResult> PublishNewWorkshopItemAsync(WorkshopItemChangeSet settings);
 
-        Task<IWorkshopItem> QueryWorkshopItemByIdAsync(int id);
+        Task<IWorkshopItem> QueryWorkshopItemByIdAsync(ulong id);
     }
 }
